experiments/requirements.txt
wget
black
isort
pandas
nltk
sklearn
sklearn_crfsuite
seqeval
attrs
matplotlib
lightgbm
scipy
tqdm
pytest
spacy
spacy-syllables
torch~=1.7.0
diskcache
sentence-transformers
clean-text
submitit
unidecode
tabulate
docopt
preprocessor
flask
joblib
[-e] git+https://github.com/awslabs/mlm-scoring.git
transformers~=3.3.1
slurmee